requirements.txt
bs4==0.0.2
selenium==4.25.0
pyautogui==0.9.54
python-dotenv==1.0.1
fake-useragent==2.0.3